Add Course, InputData, OutputData to easysnec, Refactor conftest
