+++ tests/conftest.py
@@ -1,47 +1,17 @@
 from __future__ import annotations
 
 import datetime as dt
-from dataclasses import dataclass
-from enum import Enum
-from typing import Iterable
 
-from pydantic.dataclasses import dataclass  # Type Enforcement
 import pytest
 
-
-class SuccessStatus(Enum):
-    SUCCESS = 1
-    INCOMPLETE = 2
-    MISSES = 3  # or wrong order
-    NOT_STARTED = 4
-
-
-@dataclass(frozen=True)
-class MockCourse:
-    course_name: str
-    is_score_o: bool
-    stations: list[int]
-
-
-@dataclass(frozen=True)
-class MockInputData:
-    card_id: int
-    start_time: dt.datetime
-    finish_time: dt.datetime
-    punches: list[tuple[int, dt.datetime]]
-    reading_id: str
-
-
-@dataclass(frozen=True)
-class MockOutputData:
-    course_name: str
-    success_status: SuccessStatus
-    missed_checkpoints: list[int]
+from easysnec.course import Course
+from easysnec.input_data import InputData
+from easysnec.output_data import OutputData, SuccessStatus
 
 
 @pytest.fixture
-def example_course() -> MockCourse:
-    return MockCourse(
+def example_course() -> Course:
+    return Course(
         course_name="tutorial-O",
         is_score_o=False,
         stations=[42, 43, 49],
@@ -49,7 +19,7 @@ def example_course() -> MockCourse:
 
 
 @pytest.fixture
-def example_input_success(example_course: MockCourse) -> MockInputData:
+def example_input_success(example_course: Course) -> InputData:
     start = dt.datetime(2025, 3, 14, 9, 30, 17)
     finish = dt.datetime(2025, 3, 14, 10, 22, 47)
     punches = [
@@ -57,7 +27,7 @@ def example_input_success(example_course: MockCourse) -> MockInputData:
         (43, dt.datetime(2025, 3, 14, 9, 57, 8)),
         (49, dt.datetime(2025, 3, 14, 10, 9, 44)),
     ]
-    return MockInputData(
+    return InputData(
         card_id=123,
         start_time=start,
         finish_time=finish,
@@ -67,8 +37,8 @@ def example_input_success(example_course: MockCourse) -> MockInputData:
 
 
 @pytest.fixture
-def example_output_success(example_course: MockCourse) -> MockOutputData:
-    return MockOutputData(
+def example_output_success(example_course: Course) -> OutputData:
+    return OutputData(
         course_name=example_course.course_name,
         success_status=SuccessStatus.SUCCESS,
         missed_checkpoints=[],
@@ -76,7 +46,7 @@ def example_output_success(example_course: MockCourse) -> MockOutputData:
 
 
 @pytest.fixture
-def example_input_incomplete(example_course: MockCourse) -> MockInputData:
+def example_input_incomplete(example_course: Course) -> InputData:
     start = dt.datetime(2025, 3, 14, 9, 30, 17)
     finish = dt.datetime.min  # finish time precedes start time
     punches = [
@@ -84,7 +54,7 @@ def example_input_incomplete(example_course: MockCourse) -> MockInputData:
         (43, dt.datetime(2025, 3, 14, 9, 57, 8)),
         (49, dt.datetime(2025, 3, 14, 10, 9, 44)),
     ]
-    return MockInputData(
+    return InputData(
         card_id=123,
         start_time=start,
         finish_time=finish,
@@ -94,8 +64,8 @@ def example_input_incomplete(example_course: MockCourse) -> MockInputData:
 
 
 @pytest.fixture
-def example_output_incomplete(example_course: MockCourse) -> MockOutputData:
-    return MockOutputData(
+def example_output_incomplete(example_course: Course) -> OutputData:
+    return OutputData(
         course_name=example_course.course_name,
         success_status=SuccessStatus.INCOMPLETE,
         missed_checkpoints=[],
@@ -103,14 +73,14 @@ def example_output_incomplete(example_course: MockCourse) -> MockOutputData:
 
 
 @pytest.fixture
-def example_input_misses(example_course: MockCourse) -> MockInputData:
+def example_input_misses(example_course: Course) -> InputData:
     start = dt.datetime(2025, 3, 14, 9, 30, 17)
     finish = dt.datetime(2025, 3, 14, 10, 22, 47)
     punches = [
         (43, dt.datetime(2025, 3, 14, 9, 57, 8)),
         (42, dt.datetime(2025, 3, 14, 9, 37, 2)),  # out of order
     ]
-    return MockInputData(
+    return InputData(
         card_id=123,
         start_time=start,
         finish_time=finish,
@@ -120,8 +90,8 @@ def example_input_misses(example_course: MockCourse) -> MockInputData:
 
 
 @pytest.fixture
-def example_output_misses(example_course: MockCourse) -> MockOutputData:
-    return MockOutputData(
+def example_output_misses(example_course: Course) -> OutputData:
+    return OutputData(
         course_name=example_course.course_name,
         success_status=SuccessStatus.MISSES,
         missed_checkpoints=[42, 49],
@@ -129,7 +99,7 @@ def example_output_misses(example_course: MockCourse) -> MockOutputData:
 
 
 @pytest.fixture
-def example_input_wrong_order(example_course: MockCourse) -> MockInputData:
+def example_input_wrong_order(example_course: Course) -> InputData:
     start = dt.datetime(2025, 3, 14, 9, 30, 17)
     finish = dt.datetime(2025, 3, 14, 10, 22, 47)
     punches = [
@@ -137,7 +107,7 @@ def example_input_wrong_order(example_course: MockCourse) -> MockInputData:
         (42, dt.datetime(2025, 3, 14, 9, 37, 2)),
         (43, dt.datetime(2025, 3, 14, 9, 57, 8)),
     ]
-    return MockInputData(
+    return InputData(
         card_id=123,
         start_time=start,
         finish_time=finish,
@@ -147,8 +117,8 @@ def example_input_wrong_order(example_course: MockCourse) -> MockInputData:
 
 
 @pytest.fixture
-def example_output_wrong_order(example_course: MockCourse) -> MockOutputData:
-    return MockOutputData(
+def example_output_wrong_order(example_course: Course) -> OutputData:
+    return OutputData(
         course_name=example_course.course_name,
         success_status=SuccessStatus.MISSES,
         missed_checkpoints=[49],
@@ -156,7 +126,7 @@ def example_output_wrong_order(example_course: MockCourse) -> MockOutputData:
 
 
 @pytest.fixture
-def example_input_not_started(example_course: MockCourse) -> MockInputData:
+def example_input_not_started(example_course: Course) -> InputData:
     start = dt.datetime.max  # after the finish time
     finish = dt.datetime(2025, 3, 14, 10, 22, 47)
     punches = [
@@ -164,7 +134,7 @@ def example_input_not_started(example_course: MockCourse) -> MockInputData:
         (43, dt.datetime(2025, 3, 14, 9, 57, 8)),
         (49, dt.datetime(2025, 3, 14, 10, 9, 44)),
     ]
-    return MockInputData(
+    return InputData(
         card_id=123,
         start_time=start,
         finish_time=finish,
@@ -174,9 +144,9 @@ def example_input_not_started(example_course: MockCourse) -> MockInputData:
 
 
 @pytest.fixture
-def example_output_not_started(example_course: MockCourse) -> MockOutputData:
-    return MockOutputData(
+def example_output_not_started(example_course: Course) -> OutputData:
+    return OutputData(
         course_name=example_course.course_name,
-        success_status=SuccessStatus.NOT_STARTED,
+        success_status=SuccessStatus.INCOMPLETE,
         missed_checkpoints=[],
     )
